Add files via upload
This program displays whether the given input number is a prime number or not. In this case, I defined 3 rules already to reduce some complexity such as if the number is less than or equal to 1 : it shows false, if the number is 2 : it shows true, if the number (other than 2) is completely divided by 2 : it shows false because 2 is the only even number which is prime. And then I run a (for) loop for number greater than 2. Finally, displays the final answer using if-else statements.