Add files via upload
This program shows the working of a function called as pass by value.
Pass by value: When a variable is passed by value, a copy of the variable is made. The function works on that copy, and the original variable remains unchanged.
For example, Imagine passing my xeroxed marksheet copy to the company. If they stamp or mark it, my original marksheet remains untouched. That's pass by value.